实现VideoDecoder和AudioDecoder以及测试部分，可以正常出画面，音视频同步还未完成

+++ app/src/main/java/com/chiron/playpalyer/mediac/IDecoder.kt
@@ -70,7 +70,7 @@ interface IDecoder:Runnable {
     /**
      * 获取视频长度（播放时长）
      */
-    fun getDuration():Int
+    fun getDuration():Long
 
     /**
      * 获取当前视频播放的时间戳
